+++ src/Addons/34thPRC_Overrides/data/backpacks/config_vehicles.hpp
@@ -1,0 +1,1 @@
+#include "marines\config_vehicles.hpp"