+++ src/Addons/34thPRC_Overrides/data/config_vehicles.hpp
@@ -1,1 +1,2 @@
+#include "backpacks\config_vehicles.hpp"
 #include "vehicles\config_vehicles.hpp"